Split TrackEditDialog sections into useFilters, useSamples and useTrackSettings hooks
