+++ interface/package.json
@@ -36,7 +36,7 @@
     "postcss": "https://registry.npmjs.org/postcss/-/postcss-8.4.31.tgz"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "react-scripts start GENERATE_SOURCEMAP=false",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
